final_project/requirements.txt
Flask==1.1.2
Flask-WTF==0.14.3
ibm-watson==4.4.1
python-dotenv==0.13.0
pyJWT==1.7.1
pip3 install -q PyJWT==1.7.1 flask-appbuilder==3.4.4 markupsafe==1.1.1 httpx==0.20 Flask ibm-watson python-dotenv